fix: leave parameters of tools for local MCP server
